requirements.txt
langchain
chromadb==0.4.8
onnxruntime==1.14.1
langchain_chroma==0.1.3
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
markdown2
pandas
PyPDF2
python-dotenv
python_docx
Werkzeug
django-markdown-deux
gunicorn
python_docx
exceptions
